Add Loop and Reset tile animation ops
Loop is for looping without resetting, as happens when a pattern is full. It was
previously called Repeat. Since I want to add conditions, I think 'loop' rather
than 'repeat' in the middle of a pattern is clearer.

These have no args, so the editor needed some generalisation wrt args.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@13606 7d344553-34f0-0310-a9b1-970ce8f1c3a2